Remove minimize button from user interface for projects that aren't complete.

+++ panel.manageDataset.php
@@ -147,7 +147,7 @@
 		echo "<span id='p_label_".$key."' style='color:#".$labelRgbColor.";'>\n\t\t\t\t";
 		echo "<font size='2'>".($key+1).".";
 		echo "<button id='project_delete_".$key."' type='button' onclick=\"parent.deleteProjectConfirmation('".$project."','".$key."');\">Delete</button>";
-		if (!file_exists("users/".$user."/projects/". $project ."/minimized.txt")) {
+		if (!file_exists("users/".$user."/projects/".$project."/minimized.txt") && file_exists("users/".$user."/projects/".$project."/complete.txt")) {
 			echo "<button id='project_minimize_".$key."' type='button' onclick=\"parent.minimizeProjectConfirmation('".$project."','".$key."');\">Minimize</button>";
 		}
 		echo $projectNameString;
@@ -182,7 +182,7 @@
 		}
 		echo "</font></span>\n\t\t\t\t";
 		echo "<span id='p_delete_".$key."'></span>\n";
-		if (!file_exists("users/".$user."/projects/". $project ."/minimized.txt")) {
+		if (!file_exists("users/".$user."/projects/".$project."/minimized.txt") && file_exists("users/".$user."/projects/".$project."/complete.txt")) {
 			echo "<span id='p_minimize_".$key."'></span>\n";
 		}
 		echo "\t\t\t\t";
